requirements.txt
accelerate==0.27.2
addict==2.4.0
attrs==19.3.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
config==0.5.1
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.7
diffusers==0.26.3
distro==1.4.0
distro-info==0.23+ubuntu1.1
docutils==0.16
dulwich==0.19.15
empy==3.3.2
entrypoints==0.3
fastimport==0.9.8
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gevent==24.2.1
gpg==1.13.1
greenlet==3.0.3
# Editable install with no version control (groundingdino==0.1.0)
-e /home/shizijie/experiment/Grounded-Segment-Anything-main/GroundingDINO
hf_transfer==0.1.6
html5lib==1.0.1
httplib2==0.14.0
huggingface-hub==0.21.3
hyperlink==19.0.0
idna==2.8
importlib-metadata==7.0.1
importlib_resources==6.1.2
incremental==16.10.1
Jinja2==3.1.3
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.1.5
matplotlib==3.7.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
nose==1.3.7
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
osrf-pycommon==2.0.2
packaging==23.2
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipdeptree==2.16.2
platformdirs==4.2.0
protobuf==3.6.1
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyassimp==4.1.4
pycairo==1.16.2
pycocotools==2.0.7
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2023.12.25
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
safetensors==0.4.2
scipy==1.10.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
# Editable install with no version control (segment-anything==1.0)
-e /home/shizijie/experiment/Grounded-Segment-Anything-main/segment_anything
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
ssh-import-id==5.10
supervision==0.18.0
sympy==1.12
systemd-python==234
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.2
triton==2.2.0
Twisted==18.9.0
txaio==2.10.0
typing_extensions==4.10.0
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vcstools==0.1.42
wadllib==1.3.3
webencodings==0.5.1
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
wxPython==4.0.7
xkit==0.0.0
yapf==0.40.2
zipp==3.17.0
zope.event==5.0
zope.interface==4.7.1
